requirements.txt
attrs==19.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.0
configparser==3.7.4
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
enum34==1.1.6
feather-format==0.4.0
functools32==3.2.3.post2
futures==3.2.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==5.2.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
pandas==0.24.2
pandocfilters==1.4.2
pathlib2==2.3.3
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.6.0
prompt-toolkit==1.0.16
ptyprocess==0.6.0
pyarrow==0.13.0
Pygments==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.1
qtconsole==4.4.3
scandir==1.10.0
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
subprocess32==3.5.3
terminado==0.8.2
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
